pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xml-tools"
version = "0.1.0"
description = "Inspect, model, and generate XML layouts with YAML-friendly tooling."
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
authors = [
  { name = "Willian Patrick (willian@plugfy.com)" }
]
dependencies = [
  "lxml>=4.9",
  "PyYAML>=6.0",
  "xmlschema>=2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=4.1",
  "build>=1.2",
  "twine>=4.0",
]

[project.scripts]
xml-tools = "xml_tools.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["xml_tools"]
omit = ["*/cli.py"]

[tool.coverage.report]
fail_under = 80
show_missing = true

[tool.setuptools.packages.find]
where = ["src"]
